requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
coveralls==1.5.1
dj-database-url==0.5.0
Django==2.2.2
django-cors-middleware==1.3.1
django-extensions==2.1.4
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.10.0
docopt==0.6.2
gunicorn==19.9.0
idna==2.8
isort==4.3.4
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
openapi-codec==1.3.2
psycopg2==2.7.6.1
pycodestyle==2.4.0
PyJWT==1.7.1
pylint==2.2.2
python-decouple==3.1
pytz==2018.7
requests==2.21.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.3
whitenoise==4.1.2
wrapt==1.10.11
